backend/requirements.txt
flask
flask-sqlalchemy
psycopg2-binary
flask-migrate
flask-cors
bcrypt
flask-jwt-extended
python-dotenv